back/requirements.txt
flask
flask-cors
pymongo
python-dotenv
bcrypt
pyjwt
flask-session
cryptography
pandas
Werkzeug
openpyxl
